Insanely fast tree pruning method for evaluation.
If we can prove that the absolute value of the partially evaluated polynomial
is too large for the remaining terms to reduce the polynomial to zero, we can
prune away an entire section of the tree of all possibilities in one fell
swoop. As it turns out, the plain old triangle inequality is sufficient to
compute an extremely efficient bound!

There is currently an additional fudge factor involved in the bounding to
make it less efficient but improve the aesthetic appeal of the result.